+++ pgvector/peewee/__init__.py
@@ -1,3 +1,4 @@
-from ..utils import from_db, to_db
+from ..utils import from_db, to_db, SparseVec
 from .halfvec import HalfvecField
+from .sparsevec import SparsevecField
 from .vector import VectorField