+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Added support for `halfvec` and `sparsevec` types to Psycopg 3
 - Added support for `halfvec` and `sparsevec` types to asyncpg
-- Added support for `halfvec` type to Peewee
+- Added support for `halfvec` and `sparsevec` types to Peewee
 - Added `L1Distance` for Django
 - Added `l1_distance` for SQLAlchemy, SQLModel, and Peewee